Extract Windows API helpers
