+++ binding.gyp
@@ -37,6 +37,7 @@
             }],
             ["OS=='win'", {
                 "sources": [
+                    "src/helper/windows/helper.cpp",
                     "src/worker/windows/windows_worker_platform.cpp"
                 ]
             }],
